refactor(llm): colocate provider package logic
