+++ packages/llm/package.json
@@ -16,18 +16,16 @@
     "./provider": "./src/provider.ts",
     "./providers": "./src/providers/index.ts",
     "./provider-package": "./src/provider-package.ts",
-    "./providers/amazon-bedrock": "./src/provider-packages/amazon-bedrock.ts",
-    "./providers/anthropic": "./src/provider-packages/anthropic-messages.ts",
-    "./providers/anthropic/messages": "./src/provider-packages/anthropic-messages.ts",
+    "./providers/amazon-bedrock": "./src/providers/amazon-bedrock.ts",
+    "./providers/anthropic": "./src/providers/anthropic.ts",
     "./providers/azure": "./src/providers/azure.ts",
     "./providers/cloudflare": "./src/providers/cloudflare.ts",
     "./providers/github-copilot": "./src/providers/github-copilot.ts",
     "./providers/google": "./src/providers/google.ts",
-    "./providers/openai": "./src/provider-packages/openai-responses.ts",
-    "./providers/openai/responses": "./src/provider-packages/openai-responses.ts",
-    "./providers/openai/chat": "./src/provider-packages/openai-chat.ts",
-    "./providers/openai-compatible": "./src/provider-packages/openai-compatible-chat.ts",
-    "./providers/openai-compatible/chat": "./src/provider-packages/openai-compatible-chat.ts",
+    "./providers/openai": "./src/providers/openai.ts",
+    "./providers/openai/responses": "./src/providers/openai/responses.ts",
+    "./providers/openai/chat": "./src/providers/openai/chat.ts",
+    "./providers/openai-compatible": "./src/providers/openai-compatible.ts",
     "./providers/openai-compatible-profile": "./src/providers/openai-compatible-profile.ts",
     "./providers/openrouter": "./src/providers/openrouter.ts",
     "./providers/xai": "./src/providers/xai.ts",
