--- packages/llm/src/provider-packages/shared.ts
@@ -1,8 +1,0 @@
-import type { RouteDefaultsInput } from "../route/client"
-import type { ProviderPackage } from "../provider-package"
-
-export const defaults = (settings: ProviderPackage.Settings): RouteDefaultsInput => ({
-  headers: settings.headers === undefined ? undefined : { ...settings.headers },
-  limits: settings.limits,
-  http: settings.body === undefined ? undefined : { body: { ...settings.body } },
-})
