+++ packages/llm/src/providers/openai/chat.ts
@@ -1,0 +1,2 @@
+export { chatModel as model } from "../openai"
+export type { Settings } from "../openai"
