+++ packages/llm/src/providers/openai/responses.ts
@@ -1,0 +1,2 @@
+export { model } from "../openai"
+export type { Settings } from "../openai"
